// web/src/api/generated/model/announcementStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */

export type AnnouncementStatus = typeof AnnouncementStatus[keyof typeof AnnouncementStatus];


export const AnnouncementStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  expired: 'expired',
} as const;